Format threadpool go code and add close() function.
